Consolidate app wide meta tags into component
