+++ src/components/_tests_/AppHelmet.test.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import AppHelmet from '../AppHelmet';
+
+describe('<AppHelmet />', () => {
+  it('renders valid', () => {
+    expect(React.isValidElement(<AppHelmet />)).toBeTruthy();
+  });
+});
